requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
proxy-tools==0.1.0
pycairo==1.21.0
PyGObject==3.42.1
pywebview==3.6.3
requests==2.27.1
urllib3==1.26.9
